refactor(email): move email normalization into separate file
